# services/auth/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaintAugustin Auth Service",
        "description": "Authentication, user management, and role assignment for the SaintAugustin worship platform.",
        "contact": {
            "name": "SaintAugustin",
            "email": "admin@saintaugustin.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local development (via API gateway)"
        }
    ],
    "paths": {
...
        "/imports/staug": {
            "post": {
                "tags": [
                    "Imports"
                ],
                "summary": "Import (or preview) a signed STAUG archive",
                "description": "Accepts a multipart upload of a `.staug`/`.zip` archive. The manifest signature is verified first — an invalid or unsigned archive returns 422 without writing anything. With `dry_run=true` the response lists the per-song action (create/skip/conflict). Committing is strictly non-destructive: existing songs are matched by id + title and never overwritten. Requires the `admin` role.",
                "operationId": "0994d04412dfff1dff3ebc948ac4281a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "The signed STAUG ZIP archive.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "dry_run": {
                                        "description": "If true, only preview the per-song actions; nothing is persisted.",
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preview (dry run) — per-song action and per-songbook counts.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "playlist"
                                        },
                                        "songs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "action": {
                                                        "type": "string",
                                                        "enum": [
                                                            "create",
                                                            "skip",
                                                            "conflict"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "songbooks": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Import committed — created/skipped/conflicted counts.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created": {
                                            "type": "integer"
                                        },
                                        "skipped": {
                                            "type": "integer"
                                        },
                                        "conflicted": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "songbooks": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "File missing/too large, or the archive is invalid or unsigned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "description": "Generic error response.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "description": "Laravel validation failure (HTTP 422).",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The email field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token. Obtain it from POST /api/auth/login."
            }
        }
    },
...
}